feat: use correct config from different mount points

+++ open.sh
@@ -1,3 +1,9 @@
+mount=$(realpath $1 2> /dev/null)
+if ! [[ -d $mount ]]; then
+  echo "error: $1 is not a valid mount directory."
+  exit 1
+fi
+
 xhost +localhost
 docker run --rm -it --init \
   --name spriter \
@@ -5,6 +11,7 @@ docker run --rm -it --init \
   --net=host \
   -e DISPLAY=host.docker.internal:0 \
   -v /tmp/.X11-unix:/tmp/.X11-unix:rw \
-  -v "$(pwd)":/root \
+  -v "$(dirname $BASH_SOURCE)":/root \
+  -v "$mount":/mnt \
   --privileged \
   spriter
